fix casting issues

+++ cpu/main.py
@@ -1070,22 +1070,22 @@ class CPU:
 
     def dec_zero_page(self):
         address = self.zero_page()
-        self.mem[address] -= 1
+        self.mem[address] -= np.uint8(1)
         return address, 5
 
     def dec_zero_page_x(self):
         address = self.zero_page() + self.x
-        self.mem[address] -= 1
+        self.mem[address] -= np.uint8(1)
         return address, 6
 
     def dec_absolute(self):
         address = self.absolute_address()
-        self.mem[address] -= 1
+        self.mem[address] -= np.uint8(1)
         return address, 6
 
     def dec_absolute_x(self):
         address = self.absolute_address() + self.x
-        self.mem[address] -= 1
+        self.mem[address] -= np.uint8(1)
         return address, 7
 
     def dex(self):
@@ -1108,25 +1108,25 @@ class CPU:
 
     def inc_zero_page(self):
         address = self.zero_page()
-        self.mem[address] += 1
+        self.mem[address] += np.uint8(1)
         self.set_zero_and_neg(self.mem[address])
         return address, 5
 
     def inc_zero_page_x(self):
         address = self.zero_page() + self.x
-        self.mem[address] += 1
+        self.mem[address] += np.uint8(1)
         self.set_zero_and_neg(self.mem[address])
         return address, 6
 
     def inc_absolute(self):
         address = self.absolute_address()
-        self.mem[address] += 1
+        self.mem[address] += np.uint8(1)
         self.set_zero_and_neg(self.mem[address])
         return address, 6
 
     def inc_absolute_x(self):
         address = self.absolute_address() + self.x
-        self.mem[address] += 1
+        self.mem[address] += np.uint8(1)
         self.set_zero_and_neg(self.mem[address])
         return address, 7
